Move methods to mef.cpp/map_cell.cpp/event.cpp
