+++ magic.cpp
@@ -14,6 +14,8 @@
 #include "item_db.hpp"
 #include "macro.hpp"
 #include "map.hpp"
+#include "map_cell.hpp"
+#include "mef.hpp"
 #include "trait.hpp"
 #include "variables.hpp"
 #include "wish.hpp"
@@ -3402,24 +3404,24 @@ label_2181_internal:
             }
             if (efid == 634)
             {
-                addmef(x, y, 4, 20, rnd(4) + 2, efp, cc);
+                mef_add(x, y, 4, 20, rnd(4) + 2, efp, cc);
             }
             if (efid == 455)
             {
-                addmef(x, y, 3, 19, rnd(10) + 5, efp, cc);
+                mef_add(x, y, 3, 19, rnd(10) + 5, efp, cc);
             }
             if (efid == 456)
             {
-                addmef(x, y, 5, 24, rnd(10) + 5, efp, cc);
+                mef_add(x, y, 5, 24, rnd(10) + 5, efp, cc);
                 mapitem_fire(x, y);
             }
             if (efid == 436)
             {
-                addmef(x, y, 1, 11, -1, efp * 2);
+                mef_add(x, y, 1, 11, -1, efp * 2);
             }
             if (efid == 437)
             {
-                addmef(x, y, 2, 30, 8 + rnd((15 + efp / 25)), efp);
+                mef_add(x, y, 2, 30, 8 + rnd((15 + efp / 25)), efp);
             }
         }
         break;
@@ -4222,7 +4224,7 @@ label_2181_internal:
                 }
                 if (rnd(40) == 0)
                 {
-                    addmef(dx, dy, 5, 24, rnd(4) + 3, 50);
+                    mef_add(dx, dy, 5, 24, rnd(4) + 3, 50);
                 }
                 if (map(dx, dy, 1) != 0)
                 {
@@ -4399,7 +4401,7 @@ label_2181_internal:
                 name(tc) + u8"は炎に包まれた。"s,
                 name(tc) + u8" "s + is(tc) + u8" surrounded by flames."s));
         }
-        addmef(
+        mef_add(
             cdata[tc].position.x,
             cdata[tc].position.y,
             5,
