+++ talk_unique.cpp
@@ -3,10 +3,12 @@
 #include "calc.hpp"
 #include "character.hpp"
 #include "elona.hpp"
+#include "event.hpp"
 #include "i18n.hpp"
 #include "item.hpp"
 #include "item_db.hpp"
 #include "macro.hpp"
+#include "map_cell.hpp"
 #include "variables.hpp"
 
 
@@ -24,7 +26,7 @@ talk_result_t talk_unique()
         listmax = 0;
         buff = lang(
             u8"ここまで辿り着くとはな…どうやら《混沌》は、自ら創りしネフィアの安定さえも望まぬらしい。しかし、私とてここで死ぬつもりなどないのだ。"s,
-            u8"So you've made it this far. Evidently, <Chaos> wants no poise even within their own creation...Nefia. Well, it seems they have left me no choice but to whip you!"s);
+            u8"So you've made it this far. Event_Idently, <Chaos> wants no poise even within their own creation...Nefia. Well, it seems they have left me no choice but to whip you!"s);
         tc = tc * 1 + 0;
         ELONA_APPEND_RESPONSE(0, i18n::_(u8"ui", u8"more"));
         chatesc = 1;
@@ -32,7 +34,7 @@ talk_result_t talk_unique()
         gdata_main_quest_flag = 170;
         return talk_result_t::talk_end;
     case 23:
-        if (evid() == 1)
+        if (event_id() == 1)
         {
             if (jp)
             {
@@ -2106,7 +2108,7 @@ talk_result_t talk_unique()
                 ELONA_APPEND_RESPONSE(0, i18n::_(u8"ui", u8"bye"));
                 chatesc = 1;
                 ELONA_TALK_SCENE_CUT();
-                evadd(20, tc);
+                event_add(20, tc);
                 return talk_result_t::talk_end;
             }
             if (chatval == 2)
